got BotVisualizer working with the new systems classes.  runPendulumDynamics visualizes again
