+++ drake/core/Core.h
