+++ drake/solvers/test/simpleFunctionTest.cpp
@@ -1,5 +1,5 @@
 
-#include "DrakeCore.h"
+#include "Core.h"
 #include "DrakeFunction.h"
 #include <iostream>
